[ADD] Commons 폴더 생성 및 경로정리
